log: use python-json-logger and create CustomJsonFormatter
python-json-logger is one of the most popular JSON formatter for Python logging
and it requires a custom class to fit the log spec of Stackdriver Logging.
As Python projects in this project do not share the logger module,
the custom logger module needs to be duplicated in each micro services.