Only install the Python dependencies we need
There are a lot of Python dependencies being installed for the
emailservice that aren't being used, and not all of the dependencies for
the loadgenerator are pinned.

This allows us to only specify the top-level dependencies we care about
in the requirements.in files, which are then compiled to frozen
dependencies in the requirements.txt files.