Add checkout payment path errors (#31)
We introduce an error path to the system that is fully controllable
via environment variables.

The error we introduce is that a developer has accidentally pushed code
to the payment service that uses a dev token instead of the prod token
when sending payment charge request to their external payment processor
(ButtercupPayments). The production environment performs a slow rollout
of new code, so not all pod instances have been updated to the
problematic code version. The checkoutservice has retries enabled, so if
it detects a failed call it simply retries with a backoff strategy.
Ultimately the charge call will land on a pod instance with the previous
version and succeed, but the user will experience a latency increase for
any checkout experience, which is definitely not ideal.

The hunt is on to figure out why user's are complaining about long wait
times while checking out...

Changes:
- checkoutservice now retries on any payment.charge failures. The
  number of retries and sleep time are controllable via environment
  variable.

- paymentservice now can fail requests probabilisticly based on rate
  defined in environment variable. These failures will return a 401
  due to an "invalid token" meant to mimic a dev-token that is used
  in a specific "version" of the service due to a bad code check-in.
  Any "failed token" spans are tagged using the newest version of
  code, any "successful token" spans (even if they fail for other
  reasons) are tagged using the previous version of code. The
  intermitent nature of the failure is meant to show a rollout that
  is in progress and problematic, so not all pod instances are on
  the latest versions.

- paymentservice now represents an internal service that calls an
  external payment processor (ButtercupPayments). Spans are emitted
  to represent this call, and artificial delay is introduced to
  represent the external network call.

- More verbose logs added to paymentservice to represent success and
  failure modes.