Added two manifests for E2E tests
This PR adds two more manifests that will be used for End to End tests. In particular, fixed-2clusters-manifests.yaml deploys the Online Boutique application assuming a peering between 2 clusters. In this scenario, some Pods are forced on the home clusters and others on the foreign one. Instead in fixed-3clusters-manifests.yaml, we assume to deploy the application in cluster B, peered with A and C (A and C do not have a peering session). Here, we force some Pods on the home cluster B, and others in A and C, thanks to the new Liqo node affinity capabilities.