Move YAML document separator to top of release manifests
I propose moving the YAML document separator inserted as part of `make-release-artifacts.sh` to the top of each manifest rather than the bottom.

Having the separator at the end of the manifest results in the generated release YAML files containing an empty document at the end. Whilst this is valid YAML, it is untidy, and certain YAML tools (`yq`) have been seen to handle empty document elements badly.