feat: Add support for overriding resource limits
- Allow overriding of default resource limits in deployment
manifests for microservices demo applications.
add a resources section to each container in the manifests,
with default values for CPU and memory limits.

This change will make the microservices demo more flexible and
customizable, and should make it easier for users to deploy
and scale the application in different environments.